refactor(cli): extract CLI commands to internal/cli/ package (#120)
Move all cobra command definitions out of cmd/hardbox/main.go into a
dedicated internal/cli/ package:

  internal/cli/root.go      — Execute(), newRootCmd(), applyLogLevel()
  internal/cli/apply.go     — hardbox apply subcommand
  internal/cli/audit.go     — hardbox audit subcommand
  internal/cli/rollback.go  — hardbox rollback list/apply subcommands
  internal/cli/export_test.go — white-box test shims
  internal/cli/root_test.go — ported tests (log level + command wiring)

cmd/hardbox/main.go is now a 14-line entry point that calls cli.Execute().
cmd/hardbox/main_test.go retained as compile-time smoke test.

Zero behaviour change — all existing tests pass.

Closes #120
